Allow Horizon Custom Branding
This introduces 2 variables:
``horizon_site_name`` and ``horizon_custom_uploads``.

The former needs to set with a string. It will edit the
local settings of horizon to add the cloud's name.

The latter is a dict, that contains path to source and destination
files that the deployer wants to upload.
For example, if setting the dict as followed:

horizon_custom_uploads:
  logo:
    src: '/etc/openstack_deploy/files/logo.png'
    dest: "img/logo.png"
  logosplash:
    src: '/etc/openstack_deploy/files/logo-splash.png'
    dest: "img/logo-splash.png"

You'll overwrite the img/logo.png and img/logo-splash.png files
on the horizon nodes (inside their folder
"{{ horizon_lib_dir }}/openstack_dashboard/static/dashboard/")
with the files stored on the deployment node at the src location.

Change-Id: I1a74e74968d09f6e299dbd965f9c87368c4d646c