Restart horizon when package/venvs change
Currently, upgrading a horizon package/venv does not explicitly trigger
a restart.  A restart will only occur if an explicit or implicit config
change occurs.

This commit adds several restart notifiers such that a horizon package
upgrade or venv replacement will force a restart.

Note that this issue exists in all OpenStack roles at the moment, but
the remaining roles are being addressed in their re-deploy venv
reviews.  The horizon re-deploy venv change was merged a while back
which is why we need the separate review for this.

Change-Id: I890e46bb7a2973c3582be7a37ee70730d1e3ffef