Enable new LBaaS panel in Horizon
The new LBaaS v2 panel for Horizon in Mitaka and Newton is installed
as a plugin. This patch enables the plugin as long as
`horizon_enable_neutron_lbaas` is set to True. The reference to the
older built-in LBaaS dashboard has been removed since it only works
with LBaaS v1.

Release notes are included with the patch.

Depends-On: I2d5300d4d5b02df3351ab52b1be32dd60241d34b
Change-Id: I2b1d77983598fb14fbf9ff7f23870cf767135811