Use new ansible package module to install distro packages.
From Newton onwards we're able to take advantage of
the Ansible package module, instead of conditionally
executing the apt or yum module.

Implementing this is an optimisation which we can do
in master and backport to stable/newton,
reducing the execution time.

Reference: http://docs.ansible.com/ansible/package_module.html

Change-Id: I0bce6276d390891f7cf3a1d865662f19db7b9ed0
Related-Bug: #1642654