Fix keystone_sp conditional
In patch https://review.openstack.org/#/c/309425/ we defaulted the
keystone_sp var to exist and be set to "{}". All conditionals were
changed to check if keytone_sp was "{}" rather than to check if it is
defined.

The horizon role still does a conditional to check if the var is
defined, rather than a check to see if it is "{}". This causes the
conditional to fail when the var is defaulted in user_variables, but has
no values in it.

Change-Id: Ib4cd0b168ec09a222e7a1da5672395852f3f8a1c