Ensure 0755 default mode on new directories
Some of the directory creation tasks in the horizon role did not have
default modes set. This causes problems with certain scripts, such as
the compression of static files at the end of the horizon installation.
The permissions on `/openstack/venvs` was `0700` and this prevented
the `horizon` user from running the `horizon-manage.py` script.

This patch ensures that all directories are created with `0755`
permissions by default.

Closes-Bug: 1655061
Change-Id: I11798a178df32c477ae1bb45dcf204b4b16c9d1a