Ensure dashboard plugins are found by py_pkgs
The py_pkgs plugin cannot interpret jinja, so it ignores
the current set of optional packages. This results in the
repo_build process ignoring them all, ultimately resulting
in the heat role failing because it cannot find the wheel
it needs in what the repo build has produced - nor can it
go out to get it from pypi because it's restricted by
pip.conf from doing so.

For now, until we are able to eliminate pip.conf or change
the build process to allow the role wheel builds to go out
to pypi, we ensure that py_pkgs will pick up and build
these packages by implementing them into individual
*_optional_pip_packages lists. This matches the mechanism
used in the os_neutron role.

Change-Id: Ia92eb5b1355ceceada323bd075b08729afd6280f