Adds ability to enable domain dropdown list on login page
Since Pike it's possible to set domains dropdown on horizon login screen
It might be usefull for private clouds, which allow users to see
full list of domains, which are present in openstack. As a pros it
simplifies login process for the customer, as they may select
corresponding domain from dropdown menu instead of regular input.
By default it is set to False, as this exposes list of domains
for unathorized users, which may not be good thing for public clouds.

Change-Id: I049489a552d99278c273feb6ae3cf5a4a7d4b8a1