Test using integrated build
In order to do a more complete verification of any patches,
we add a full containerised OpenStack deployment to do the
functional testing using the integrated repo. This replaces
the previous functional test mechanism.

A containerised build is used due to the clash between nginx
(repo server) and apache (horizon) on the same ports.

Any additional role tests are left as-is. They will require
some extra implementation in the integrated build before they
can be transferred.

Depends-On: https://review.openstack.org/647840
Depends-On: https://review.openstack.org/648502
Depends-On: https://review.openstack.org/648551
Change-Id: I19a28191ff3c068ee8134d938656bb476e5675da