Place policy in correct folder
This patch registers the policy paths in a seperate task in order
to be able to place them in the correct folder, alongside the other
policy files so that Horizon can load properly.

Without this, the heat_policy.json file doesn't get linked in the
config location and the dashboard fails to load due to a missing
policy file.

Change-Id: Ic1e7e3ae5bb2f8d56602209c185a1ed025c0fdfa