Display name in IDP dropdown
If an operator would like a different display name
on the horizon Identity Provider dropdown to the one used
for their IDP in the backend they may specify a display_name
in their entry within the keystone_sp.trusted_idp_list.
Otherwise this continues to default to the IDPs name as usual.

Change-Id: Iba8c6e6b66480cdd575f75162c9be80e84ec868b