Fix race condition in compression of static files
At present, the collection and compression of static file is
carried out asynchronously to speed up execution. If compression
starts before collection has completed it can result in incomplete
results which cause the Horizon front end to produce 500 errors.

This patch switches the collection of static files back to a
synchronous task before triggering the slower compression as an
asynchronous task to maintain most of the speed-up benefit.

Change-Id: If9df63b2ef0031c91303d3fcb56474cbf7d648f9